v0.2.0-alpha.5 (#28)
* fix(core): update `peerDependencies` to include firebase-tools v7
* fix(buildRtdbCommand): include preceding slash if it doesn't already exist on action path
* fix(createTestEnvFile): move `createTestEnvFile` to its own file in src
* fix(run): only include `envName` if argument is provided
* fix(commands): switch `cy.logout` to use `onAuthStateChanged` - #20
