+++ cmds/run.js
@@ -1,32 +1,33 @@
-/**
- * run commander component
- * To use add require('../cmds/run.js')(program) to your commander.js based node executable before program.parse
- */
-
 const chalk = require('chalk');
 const { runCommand } = require('../lib/utils');
 const logger = require('../lib/logger');
+const createTestEnvFile = require('../lib/createTestEnvFile');
 
 /**
  * @name run
- * @description Deploy to Firebase only on build branches (master, stage, prod)
+ * @description Build test configuration file then run cypress run command
  * @param {String} envName
  */
 module.exports = function run(program) {
   program
     .command('run [envName]')
-    .description(
-      'Build configuration file containing a token for authorizing a firebase instance'
-    )
-    .action((envArg) => {
-      const envName = typeof envArg === 'string' ? envArg : 'local';
-      return runCommand({ command: `cypress-firebase createTestEnvFile ${envName}` })
+    .description('Build test configuration file then run cypress run command')
+    .action(envArg => {
+      return createTestEnvFile(envArg)
         .then(() => {
-          logger.info(`Starting test run for environment: ${chalk.cyan(envName)}`);
-          return runCommand({ command: 'npx', args: ['cypress', 'run', '--env', `envName=${envName}`] });
+          logger.info(
+            `Starting test run for environment: ${chalk.cyan(envArg)}`
+          );
+          const defaultArgs = ['cypress', 'run'];
+          return runCommand({
+            command: 'npx',
+            args: envArg
+              ? defaultArgs.concat(['--env', `envName=${envArg}`])
+              : defaultArgs
+          });
         })
         .then(() => process.exit(0))
-        .catch((err) => {
+        .catch(err => {
           logger.error(`Run could not be completed:\n${err.message}`);
           process.exit(1);
           return Promise.reject(err);
