+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cypress-firebase",
-  "version": "0.2.0-alpha.4",
+  "version": "0.2.0-alpha.5",
   "description": "Utilities to help testing Firebase projects with Cypress.",
   "main": "lib/index.js",
   "module": "lib/index.js",
@@ -21,13 +21,13 @@
     "cypress-firebase": "./bin/cypress-firebase"
   },
   "peerDependencies": {
-    "firebase-tools": "^6"
+    "firebase-tools": "^6 || ^7"
   },
   "dependencies": {
     "chalk": "^2.4.2",
     "commander": "^2.19.0",
     "figures": "^3.0.0",
-    "firebase-admin": "^8.0.0",
+    "firebase-admin": "^8.1.0",
     "firebase-tools-extra": "0.1.0-alpha.2",
     "lodash": "^4.17.11"
   },
